Remove global variables, pass context across all components and improve error handling #603
##### ISSUE TYPE
 - Feature Pull Request

##### SUMMARY
- Use envconfig
- Remove global variables related to Kubernetes, logger, and filters
- Pass context across all components - support graceful shutdown of all BotKube components
- Improve error handling (return errors in almost all places instead of ignoring/logging them)

This is just a start, as later (during new features development) we should refactor each component/package in isolation. But after this huge PR it should be possible to do it.

Fixes #220 

##### TODO

After first review of this PR if the new filter engine approach is accepted: 
- [x] Update documentation (https://www.botkube.io/filters/) PR: https://github.com/infracloudio/botkube-docs/pull/81

##### TESTING

Tested manually with Mattermost, Discord and Slack.
You can use the following image:`pkosiec/botkube:remove-global-vars-v2`